--- a/src/ImageSharp/Processing/Extensions/DiffuseExtensions.cs
+++ b/src/ImageSharp/Processing/Extensions/DiffuseExtensions.cs
@@ -1,4 +1,4 @@
-﻿// Copyright (c) Six Labors and contributors.
+// Copyright (c) Six Labors and contributors.
 // Licensed under the Apache License, Version 2.0.
 
 using System;