--- a/src/ImageSharp/Processing/Processors/Dithering/AtkinsonDiffuser.cs
+++ b/src/ImageSharp/Processing/Processors/Dithering/AtkinsonDiffuser.cs
@@ -1,4 +1,4 @@
-﻿// Copyright (c) Six Labors and contributors.
+// Copyright (c) Six Labors and contributors.
 // Licensed under the Apache License, Version 2.0.
 
 using SixLabors.ImageSharp.Primitives;
@@ -11,22 +11,24 @@ namespace SixLabors.ImageSharp.Processing.Processors.Dithering
     /// </summary>
     public sealed class AtkinsonDiffuser : ErrorDiffuser
     {
+        private const float Divisor = 8F;
+
         /// <summary>
         /// The diffusion matrix
         /// </summary>
         private static readonly DenseMatrix<float> AtkinsonMatrix =
             new float[,]
             {
-               { 0, 0, 1, 1 },
-               { 1, 1, 1, 0 },
-               { 0, 1, 0, 0 }
+               { 0, 0, 1 / Divisor, 1 / Divisor },
+               { 1 / Divisor, 1 / Divisor, 1 / Divisor, 0 },
+               { 0, 1 / Divisor, 0, 0 }
             };
 
         /// <summary>
         /// Initializes a new instance of the <see cref="AtkinsonDiffuser"/> class.
         /// </summary>
         public AtkinsonDiffuser()
-            : base(AtkinsonMatrix, 8)
+            : base(AtkinsonMatrix)
         {
         }
     }