--- a/src/ImageSharp/Processing/Processors/Dithering/BurksDiffuser.cs
+++ b/src/ImageSharp/Processing/Processors/Dithering/BurksDiffuser.cs
@@ -1,4 +1,4 @@
-﻿// Copyright (c) Six Labors and contributors.
+// Copyright (c) Six Labors and contributors.
 // Licensed under the Apache License, Version 2.0.
 
 using SixLabors.ImageSharp.Primitives;
@@ -11,21 +11,23 @@ namespace SixLabors.ImageSharp.Processing.Processors.Dithering
     /// </summary>
     public sealed class BurksDiffuser : ErrorDiffuser
     {
+        private const float Divisor = 32F;
+
         /// <summary>
         /// The diffusion matrix
         /// </summary>
         private static readonly DenseMatrix<float> BurksMatrix =
             new float[,]
             {
-                { 0, 0, 0, 8, 4 },
-                { 2, 4, 8, 4, 2 }
+                { 0, 0, 0, 8 / Divisor, 4 / Divisor },
+                { 2 / Divisor, 4 / Divisor, 8 / Divisor, 4 / Divisor, 2 / Divisor }
             };
 
         /// <summary>
         /// Initializes a new instance of the <see cref="BurksDiffuser"/> class.
         /// </summary>
         public BurksDiffuser()
-            : base(BurksMatrix, 32)
+            : base(BurksMatrix)
         {
         }
     }