--- a/src/ImageSharp/Processing/Processors/Dithering/ErrorDiffuser.cs
+++ b/src/ImageSharp/Processing/Processors/Dithering/ErrorDiffuser.cs
@@ -1,4 +1,4 @@
-﻿// Copyright (c) Six Labors and contributors.
+// Copyright (c) Six Labors and contributors.
 // Licensed under the Apache License, Version 2.0.
 
 using System;
@@ -15,61 +15,32 @@ namespace SixLabors.ImageSharp.Processing.Processors.Dithering
     /// </summary>
     public abstract class ErrorDiffuser : IErrorDiffuser
     {
-        /// <summary>
-        /// The vector to perform division.
-        /// </summary>
-        private readonly Vector4 divisorVector;
-
-        /// <summary>
-        /// The matrix width.
-        /// </summary>
-        private readonly int matrixHeight;
-
-        /// <summary>
-        /// The matrix height.
-        /// </summary>
-        private readonly int matrixWidth;
-
-        /// <summary>
-        /// The offset at which to start the dithering operation.
-        /// </summary>
         private readonly int startingOffset;
-
-        /// <summary>
-        /// The diffusion matrix.
-        /// </summary>
         private readonly DenseMatrix<float> matrix;
 
         /// <summary>
         /// Initializes a new instance of the <see cref="ErrorDiffuser"/> class.
         /// </summary>
         /// <param name="matrix">The dithering matrix.</param>
-        /// <param name="divisor">The divisor.</param>
-        internal ErrorDiffuser(in DenseMatrix<float> matrix, byte divisor)
+        internal ErrorDiffuser(in DenseMatrix<float> matrix)
         {
-            Guard.MustBeGreaterThan(divisor, 0, nameof(divisor));
-
-            this.matrix = matrix;
-            this.matrixWidth = this.matrix.Columns;
-            this.matrixHeight = this.matrix.Rows;
-            this.divisorVector = new Vector4(divisor);
-
             this.startingOffset = 0;
-            for (int i = 0; i < this.matrixWidth; i++)
+
+            for (int col = 0; col < matrix.Columns; col++)
             {
-                // Good to disable here as we are not comparing mathematical output.
-                // ReSharper disable once CompareOfFloatsByEqualityOperator
-                if (matrix[0, i] != 0)
+                if (matrix[0, col] != 0)
                 {
-                    this.startingOffset = (byte)(i - 1);
+                    this.startingOffset = col - 1;
                     break;
                 }
             }
+
+            this.matrix = matrix;
         }
 
         /// <inheritdoc />
-        [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        public void Dither<TPixel>(ImageFrame<TPixel> image, TPixel source, TPixel transformed, int x, int y, int minX, int minY, int maxX, int maxY)
+        [MethodImpl(InliningOptions.ShortMethod)]
+        public void Dither<TPixel>(ImageFrame<TPixel> image, TPixel source, TPixel transformed, int x, int y, int minX, int maxX, int maxY)
             where TPixel : struct, IPixel<TPixel>
         {
             image[x, y] = transformed;
@@ -82,42 +53,37 @@ namespace SixLabors.ImageSharp.Processing.Processors.Dithering
 
             // Calculate the error
             Vector4 error = source.ToVector4() - transformed.ToVector4();
-            this.DoDither(image, x, y, minX, minY, maxX, maxY, error);
+            this.DoDither(image, x, y, minX, maxX, maxY, error);
         }
 
-        [MethodImpl(MethodImplOptions.NoInlining)]
-        private void DoDither<TPixel>(ImageFrame<TPixel> image, int x, int y, int minX, int minY, int maxX, int maxY, Vector4 error)
+        [MethodImpl(InliningOptions.ShortMethod)]
+        private void DoDither<TPixel>(ImageFrame<TPixel> image, int x, int y, int minX, int maxX, int maxY, Vector4 error)
             where TPixel : struct, IPixel<TPixel>
         {
+            int offset = this.startingOffset;
+            DenseMatrix<float> matrix = this.matrix;
+
             // Loop through and distribute the error amongst neighboring pixels.
-            for (int row = 0; row < this.matrixHeight; row++)
+            for (int row = 0, targetY = y + row; row < matrix.Rows && targetY < maxY; row++)
             {
-                int matrixY = y + row;
-                if (matrixY > minY && matrixY < maxY)
+                Span<TPixel> rowSpan = image.GetPixelRowSpan(targetY);
+
+                for (int col = 0; col < matrix.Columns; col++)
                 {
-                    Span<TPixel> rowSpan = image.GetPixelRowSpan(matrixY);
-
-                    for (int col = 0; col < this.matrixWidth; col++)
+                    int targetX = x + (col - offset);
+                    if (targetX > minX && targetX < maxX)
                     {
-                        int matrixX = x + (col - this.startingOffset);
-
-                        if (matrixX > minX && matrixX < maxX)
+                        float coefficient = matrix[row, col];
+                        if (coefficient == 0)
                         {
-                            float coefficient = this.matrix[row, col];
-
-                            // Good to disable here as we are not comparing mathematical output.
-                            // ReSharper disable once CompareOfFloatsByEqualityOperator
-                            if (coefficient == 0)
-                            {
-                                continue;
-                            }
-
-                            ref TPixel pixel = ref rowSpan[matrixX];
-                            var offsetColor = pixel.ToVector4();
-
-                            Vector4 result = ((error * coefficient) / this.divisorVector) + offsetColor;
-                            pixel.FromVector4(result);
+                            continue;
                         }
+
+                        ref TPixel pixel = ref rowSpan[targetX];
+                        var offsetColor = pixel.ToVector4();
+
+                        Vector4 result = (error * coefficient) + offsetColor;
+                        pixel.FromVector4(result);
                     }
                 }
             }