--- a/src/ImageSharp/Processing/Processors/Dithering/FloydSteinbergDiffuser.cs
+++ b/src/ImageSharp/Processing/Processors/Dithering/FloydSteinbergDiffuser.cs
@@ -1,4 +1,4 @@
-﻿// Copyright (c) Six Labors and contributors.
+// Copyright (c) Six Labors and contributors.
 // Licensed under the Apache License, Version 2.0.
 
 using SixLabors.ImageSharp.Primitives;
@@ -11,21 +11,23 @@ namespace SixLabors.ImageSharp.Processing.Processors.Dithering
     /// </summary>
     public sealed class FloydSteinbergDiffuser : ErrorDiffuser
     {
+        private const float Divisor = 16F;
+
         /// <summary>
         /// The diffusion matrix
         /// </summary>
         private static readonly DenseMatrix<float> FloydSteinbergMatrix =
             new float[,]
             {
-                { 0, 0, 7 },
-                { 3, 5, 1 }
+                { 0, 0, 7 / Divisor },
+                { 3 / Divisor, 5 / Divisor, 1 / Divisor }
             };
 
         /// <summary>
         /// Initializes a new instance of the <see cref="FloydSteinbergDiffuser"/> class.
         /// </summary>
         public FloydSteinbergDiffuser()
-            : base(FloydSteinbergMatrix, 16)
+            : base(FloydSteinbergMatrix)
         {
         }
     }