--- a/src/ImageSharp/Processing/Processors/Dithering/IErrorDiffuser.cs
+++ b/src/ImageSharp/Processing/Processors/Dithering/IErrorDiffuser.cs
@@ -1,4 +1,4 @@
-﻿// Copyright (c) Six Labors and contributors.
+// Copyright (c) Six Labors and contributors.
 // Licensed under the Apache License, Version 2.0.
 
 using SixLabors.ImageSharp.PixelFormats;
@@ -19,11 +19,10 @@ namespace SixLabors.ImageSharp.Processing.Processors.Dithering
         /// <param name="x">The column index.</param>
         /// <param name="y">The row index.</param>
         /// <param name="minX">The minimum column value.</param>
-        /// <param name="minY">The minimum row value.</param>
         /// <param name="maxX">The maximum column value.</param>
         /// <param name="maxY">The maximum row value.</param>
         /// <typeparam name="TPixel">The pixel format.</typeparam>
-        void Dither<TPixel>(ImageFrame<TPixel> image, TPixel source, TPixel transformed, int x, int y, int minX, int minY, int maxX, int maxY)
+        void Dither<TPixel>(ImageFrame<TPixel> image, TPixel source, TPixel transformed, int x, int y, int minX, int maxX, int maxY)
             where TPixel : struct, IPixel<TPixel>;
     }
 }