--- a/src/ImageSharp/Processing/Processors/Dithering/JarvisJudiceNinkeDiffuser.cs
+++ b/src/ImageSharp/Processing/Processors/Dithering/JarvisJudiceNinkeDiffuser.cs
@@ -1,4 +1,4 @@
-﻿// Copyright (c) Six Labors and contributors.
+// Copyright (c) Six Labors and contributors.
 // Licensed under the Apache License, Version 2.0.
 
 using SixLabors.ImageSharp.Primitives;
@@ -11,22 +11,24 @@ namespace SixLabors.ImageSharp.Processing.Processors.Dithering
     /// </summary>
     public sealed class JarvisJudiceNinkeDiffuser : ErrorDiffuser
     {
+        private const float Divisor = 48F;
+
         /// <summary>
         /// The diffusion matrix
         /// </summary>
         private static readonly DenseMatrix<float> JarvisJudiceNinkeMatrix =
             new float[,]
             {
-                { 0, 0, 0, 7, 5 },
-                { 3, 5, 7, 5, 3 },
-                { 1, 3, 5, 3, 1 }
+                { 0, 0, 0, 7 / Divisor, 5 / Divisor },
+                { 3 / Divisor, 5 / Divisor, 7 / Divisor, 5 / Divisor, 3 / Divisor },
+                { 1 / Divisor, 3 / Divisor, 5 / Divisor, 3 / Divisor, 1 / Divisor }
             };
 
         /// <summary>
         /// Initializes a new instance of the <see cref="JarvisJudiceNinkeDiffuser"/> class.
         /// </summary>
         public JarvisJudiceNinkeDiffuser()
-            : base(JarvisJudiceNinkeMatrix, 48)
+            : base(JarvisJudiceNinkeMatrix)
         {
         }
     }