--- a/src/ImageSharp/Processing/Processors/Dithering/Sierra3Diffuser.cs
+++ b/src/ImageSharp/Processing/Processors/Dithering/Sierra3Diffuser.cs
@@ -1,4 +1,4 @@
-﻿// Copyright (c) Six Labors and contributors.
+// Copyright (c) Six Labors and contributors.
 // Licensed under the Apache License, Version 2.0.
 
 using SixLabors.ImageSharp.Primitives;
@@ -11,22 +11,24 @@ namespace SixLabors.ImageSharp.Processing.Processors.Dithering
     /// </summary>
     public sealed class Sierra3Diffuser : ErrorDiffuser
     {
+        private const float Divisor = 32F;
+
         /// <summary>
         /// The diffusion matrix
         /// </summary>
         private static readonly DenseMatrix<float> Sierra3Matrix =
             new float[,]
             {
-               { 0, 0, 0, 5, 3 },
-               { 2, 4, 5, 4, 2 },
-               { 0, 2, 3, 2, 0 }
+               { 0, 0, 0, 5 / Divisor, 3 / Divisor },
+               { 2 / Divisor, 4 / Divisor, 5 / Divisor, 4 / Divisor, 2 / Divisor },
+               { 0, 2 / Divisor, 3 / Divisor, 2 / Divisor, 0 }
             };
 
         /// <summary>
         /// Initializes a new instance of the <see cref="Sierra3Diffuser"/> class.
         /// </summary>
         public Sierra3Diffuser()
-            : base(Sierra3Matrix, 32)
+            : base(Sierra3Matrix)
         {
         }
     }