… src/ImageSharp/Processing/Processors/Dithering/SierraLiteDiffuser.cs
… src/ImageSharp/Processing/Processors/Dithering/StevensonArceDiffuser.cs
… src/ImageSharp/Processing/Processors/Dithering/StuckiDiffuser.cs
… src/ImageSharp/Processing/Processors/Quantization/FrameQuantizer{TPixel}.cs
… src/ImageSharp/Processing/Processors/Quantization/OctreeFrameQuantizer{TPixel}.cs
… src/ImageSharp/Processing/Processors/Quantization/PaletteFrameQuantizer{TPixel}.cs
… src/ImageSharp/Processing/Processors/Quantization/WuFrameQuantizer{TPixel}.cs
… tests/ImageSharp.Benchmarks/Samplers/Diffuse.cs