### Prerequisites

- [x] I have written a descriptive pull-request title
- [x] I have verified that there are no overlapping [pull-requests](https://github.com/SixLabors/ImageSharp/pulls) open
- [x] I have verified that I am following matches the existing coding patterns and practice as demonstrated in the repository. These follow strict Stylecop rules :cop:.
- [x] I have provided test coverage for my change (where applicable)

### Description
- Optimizes the error diffusion algorithm providing a minor speedup and memory reduction. Touches #757
- Fixes an issue where too much error was being introduced.

### Benchmarks
``` bash
BenchmarkDotNet=v0.11.5, OS=Windows 10.0.18362
Intel Core i7-8650U CPU 1.90GHz (Kaby Lake R), 1 CPU, 8 logical and 4 physical cores
.NET Core SDK=3.0.100
  [Host] : .NET Core 2.1.13 (CoreCLR 4.6.28008.01, CoreFX 4.6.28008.01), 64bit RyuJIT
  Clr    : .NET Framework 4.7.2 (CLR 4.0.30319.42000), 64bit RyuJIT-v4.8.4018.0
  Core   : .NET Core 2.1.13 (CoreCLR 4.6.28008.01, CoreFX 4.6.28008.01), 64bit RyuJIT

IterationCount=3  LaunchCount=1  WarmupCount=3
```
**Before**

|    Method |  Job | Runtime |      Mean |    Error |   StdDev | Gen 0 | Gen 1 | Gen 2 | Allocated |
|---------- |----- |-------- |----------:|---------:|---------:|------:|------:|------:|----------:|
| DoDiffuse |  Clr |     Clr | 129.58 ms | 24.60 ms | 1.349 ms |     - |     - |     - |      6 KB |
| DoDiffuse | Core |    Core |  92.63 ms | 89.78 ms | 4.921 ms |     - |     - |     - |   4.58 KB |

**After**

|    Method |  Job | Runtime |      Mean |     Error |    StdDev | Gen 0 | Gen 1 | Gen 2 | Allocated |
|---------- |----- |-------- |----------:|----------:|----------:|------:|------:|------:|----------:|
| DoDiffuse |  Clr |     Clr | 124.93 ms | 33.297 ms | 1.8251 ms |     - |     - |     - |      2 KB |
| DoDiffuse | Core |    Core |  89.63 ms |  9.895 ms | 0.5424 ms |     - |     - |     - |   1.91 KB |

### Output 

**Before** 

![DiffusionFilter_WorksWithAllErrorDiffusers_Bike_FloydSteinbergDiffuser Before](https://user-images.githubusercontent.com/385879/67575765-fdf7d000-f788-11e9-9102-83a42218a473.png)

**After**

![DiffusionFilter_WorksWithAllErrorDiffusers_Bike_FloydSteinbergDiffuser](https://user-images.githubusercontent.com/385879/67620190-e0da0480-f84f-11e9-95a9-a7150c784b4d.png)

Additionally here's the output quality of our Wu Quantizer saving a png with the new diffusion

**Original** 73.8kb

![EHuYxnRUcAAsrCY-in](https://user-images.githubusercontent.com/385879/67576317-10263e00-f78a-11e9-908c-04c95d846e17.png)

**Quantized**  21.6kb
![EHuZAfsU0AAgcs_](https://user-images.githubusercontent.com/385879/67576318-10263e00-f78a-11e9-851d-ae12d5bbf199.png)
<!-- Thanks for contributing to ImageSharp! -->